rename files to kebab-case and exported items to match the file names
