+++ src/components/App/App.tsx
@@ -1,12 +1,12 @@
-import MainPage from '@/pages/Main/MainPage';
+import Main from '@/pages/main/main';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import '@/config';
 import { AppRoute } from '@/config';
-import LoginPage from '@/pages/Login/LoginPage';
+import Login from '@/pages/login/login';
 import PageNotFound from '@/pages/404/404';
-import FavoritesPage from '@/pages/Favorites/FavoritesPage';
-import OfferPage from '@/pages/Offer/OfferPage';
-import PrivateRoute from '../PrivateRoute/PrivateRoute';
+import Favorites from '@/pages/favorites/favorites';
+import Offer from '@/pages/offer/offer';
+import PrivateRoute from '@/components/private-route/private-route';
 
 type AppProps = {
   placesCount: number;
@@ -18,15 +18,15 @@ export default function App(props: AppProps) {
       <Routes>
         <Route
           path={AppRoute.Main}
-          element={<MainPage placesCount={props.placesCount} />}
+          element={<Main placesCount={props.placesCount} />}
         />
-        <Route path={AppRoute.Login} element={<LoginPage />} />
-        <Route path={AppRoute.Offer} element={<OfferPage />} />
+        <Route path={AppRoute.Login} element={<Login />} />
+        <Route path={AppRoute.Offer} element={<Offer />} />
         <Route
           path={AppRoute.Favorites}
           element={
             <PrivateRoute auth={false}>
-              <FavoritesPage />
+              <Favorites />
             </PrivateRoute>
           }
         />
