+++ src/index.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import App from '@/components/App/App';
+import App from '@/components/app/app';
 
 const root = document.getElementById('root')!;
