+++ src/pages/login/login.tsx
@@ -1,4 +1,4 @@
-export default function LoginPage(): JSX.Element {
+export default function Login(): JSX.Element {
   return (
     <div className='page page--gray page--login'>
       <header className='header'>
