+++ src/pages/offer-not-logged-in/offer-not-logged-in.tsx
@@ -1,4 +1,4 @@
-export default function OfferPage(): JSX.Element {
+export default function OfferNotLoggedIn(): JSX.Element {
   return (
     <div className='page'>
       <header className='header'>
@@ -23,15 +23,7 @@ export default function OfferPage(): JSX.Element {
                     href='#'
                   >
                     <div className='header__avatar-wrapper user__avatar-wrapper'></div>
-                    <span className='header__user-name user__name'>
-                      Oliver.conner@gmail.com
-                    </span>
-                    <span className='header__favorite-count'>3</span>
-                  </a>
-                </li>
-                <li className='header__nav-item'>
-                  <a className='header__nav-link' href='#'>
-                    <span className='header__signout'>Sign out</span>
+                    <span className='header__login'>Sign in</span>
                   </a>
                 </li>
               </ul>
@@ -205,121 +197,6 @@ export default function OfferPage(): JSX.Element {
                     </div>
                   </li>
                 </ul>
-                <form className='reviews__form form' action='#' method='post'>
-                  <label
-                    className='reviews__label form__label'
-                    htmlFor='review'
-                  >
-                    Your review
-                  </label>
-                  <div className='reviews__rating-form form__rating'>
-                    <input
-                      className='form__rating-input visually-hidden'
-                      name='rating'
-                      value='5'
-                      id='5-stars'
-                      type='radio'
-                    />
-                    <label
-                      htmlFor='5-stars'
-                      className='reviews__rating-label form__rating-label'
-                      title='perfect'
-                    >
-                      <svg className='form__star-image' width='37' height='33'>
-                        <use xlinkHref='#icon-star'></use>
-                      </svg>
-                    </label>
-
-                    <input
-                      className='form__rating-input visually-hidden'
-                      name='rating'
-                      value='4'
-                      id='4-stars'
-                      type='radio'
-                    />
-                    <label
-                      htmlFor='4-stars'
-                      className='reviews__rating-label form__rating-label'
-                      title='good'
-                    >
-                      <svg className='form__star-image' width='37' height='33'>
-                        <use xlinkHref='#icon-star'></use>
-                      </svg>
-                    </label>
-
-                    <input
-                      className='form__rating-input visually-hidden'
-                      name='rating'
-                      value='3'
-                      id='3-stars'
-                      type='radio'
-                    />
-                    <label
-                      htmlFor='3-stars'
-                      className='reviews__rating-label form__rating-label'
-                      title='not bad'
-                    >
-                      <svg className='form__star-image' width='37' height='33'>
-                        <use xlinkHref='#icon-star'></use>
-                      </svg>
-                    </label>
-
-                    <input
-                      className='form__rating-input visually-hidden'
-                      name='rating'
-                      value='2'
-                      id='2-stars'
-                      type='radio'
-                    />
-                    <label
-                      htmlFor='2-stars'
-                      className='reviews__rating-label form__rating-label'
-                      title='badly'
-                    >
-                      <svg className='form__star-image' width='37' height='33'>
-                        <use xlinkHref='#icon-star'></use>
-                      </svg>
-                    </label>
-
-                    <input
-                      className='form__rating-input visually-hidden'
-                      name='rating'
-                      value='1'
-                      id='1-star'
-                      type='radio'
-                    />
-                    <label
-                      htmlFor='1-star'
-                      className='reviews__rating-label form__rating-label'
-                      title='terribly'
-                    >
-                      <svg className='form__star-image' width='37' height='33'>
-                        <use xlinkHref='#icon-star'></use>
-                      </svg>
-                    </label>
-                  </div>
-                  <textarea
-                    className='reviews__textarea form__textarea'
-                    id='review'
-                    name='review'
-                    placeholder='Tell how was your stay, what you like and what can be improved'
-                  />
-                  <div className='reviews__button-wrapper'>
-                    <p className='reviews__help'>
-                      To submit review please make sure to set{' '}
-                      <span className='reviews__star'>rating</span> and describe
-                      your stay with at least{' '}
-                      <b className='reviews__text-amount'>50 characters</b>.
-                    </p>
-                    <button
-                      className='reviews__submit form__submit button'
-                      type='submit'
-                      disabled
-                    >
-                      Submit
-                    </button>
-                  </div>
-                </form>
               </section>
             </div>
           </div>
